Seperated AbstractTeensyCAN to a new file
